separated file_version and rnote_version, restructuring, changed magic number
