+++ crates/rnote-engine/src/meson.build
@@ -27,8 +27,8 @@ rnote_engine_sources = files(
     'fileformats/rnoteformat/mod.rs',
     'fileformats/rnoteformat/bcursor.rs',
     'fileformats/rnoteformat/compression.rs',
-    'fileformats/rnoteformat/maj0min14.rs',
     'fileformats/rnoteformat/prelude.rs',
+    'fileformats/rnoteformat/v1.rs',
     'fileformats/xoppformat.rs',
     'lib.rs',
     'pens/brush.rs',
